Use IHostedService for device access management
